pyproject.toml
[project]
name = "legal-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.128.8",
    "numpy<2.0",
    "openai>=2.20.0",
    "openpyxl>=3.1.5",
    "pandas<2.2",
    "pdf2image>=1.17.0",
    "pillow>=11.3.0",
    "pypdf>=6.7.0",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.39.0",
]
